[BUGFIX] Streamline base path handling for relative paths
When relative paths are passed and resolved using
`FilesystemHelper::resolveRelativePath()`, the base path determination
logic was currently not always equal. Sometimes, the current working
directory was used, whereas other code parts used the project root path.

In order to streamline the logic, the path of a project's composer.json
is now always used as base path, unless the current execution context is
not a running PHAR file. In this case, the current working directory is
still used as fallback.